[frontend] header changes
